+++ src/engine/Timer.h
@@ -120,7 +120,7 @@ class Timer {
         void nextStep();
 
         /// Compute the interpolation factor
-        double computeInterpolationFactor();
+        decimal computeInterpolationFactor();
 };
 
 // Return the timestep of the physics engine
@@ -189,8 +189,8 @@ inline void Timer::nextStep() {
 }
 
 // Compute the interpolation factor
-inline double Timer::computeInterpolationFactor() {
-    return (mAccumulator / mTimeStep);
+inline decimal Timer::computeInterpolationFactor() {
+    return (decimal(mAccumulator / mTimeStep));
 }
 
 // Compute the time since the last update() call and add it to the accumulator
